Replace iterable test with string, invert results.
Apparently Jinja treats strings as iterables, which yes, technically in
python they are, but still. (I see now that there's also a sequence
test, but I don't see how it's any different/better for this than
iterable.) Anyway, it works just as well to just test for if it's a
string and handle lists as the else. Fine.